Renamed bootstrap to runway, moved init_logging to runway
